patient-care-app/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==44.0.0
python-dotenv==1.0.1
python-multipart==0.0.17
